backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
pydantic==2.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
aiohttp==3.9.3
